Update README.md
My current academic career path is to obtain an undergraduate certificate in machine Learning and design techniques an an Associates in science in Engineering Technology from DeVry University. I will be entering into a Dual Bachelors program in Artificial Intelligence and Advancing computer science in the fall of '25. The Fall of '27 I hope to to Enter graduate school at UC Berkely and Obtain a Masters in Molecular science in Software Engineering. Then I hope to enter into a PHD Program to obtain my PHD in AI. I would like to gain real world experience on this career path to Go simualtaneously with my Academic credentials to become a solid asset to a future Tech Company.